Refactor ToggleShoutCommand and ToggleWhisperCommand from badfaction to chat package
